Correct to set cookie.httpOnly option to false
Even if the httpOnly option is set to false, it is set to true.
The logical OR (||) operator (logical disjunction) for a set of operands is true if and only if one or more of its operands is true.
So changed to use Nullish coalescing operator(??)

Issue: https://github.com/hoangvvo/next-session/issues/353

See also
  - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Logical_OR